Re-addressed the issue of BlockMapData's Directory attribute, and the fact that Directory was holding a Map attribute. The original problem was free-memory-read errors that were occurring when the Directory's Map attribute was destroyed and then another Map (which was referencing the same BlockMapData object) would call Directory::GetDirectoryEntries, causing Directory to reference the Map that no longer existed. Now, Directory::GetDirectoryEntries accepts a Map argument rather than depending on the Map that was stored as an attribute when Directory was constructed.
Map objects now create the Directory object the first time it is needed, but
don't destroy it until the BlockMapData object is destroyed. This avoids
construction overhead for cases where the Directory is used repeatedly.

Ran all epetra tests in serial and parallel on linux, and ran the
BlockMap and CrsMatrix tests with purify on solaris in parallel.