Two major commits:
1)  NOX Multivector support:  Added an abstract NOX multivector
(NOX::Abstract::MultiVector) and a default implementation using an
array of NOX::Abstract::Vector's (NOX::MultiVector).  Also added an
Epetra implementation (NOX::Epetra::MultiVector).  The Teuchos
SerialDenseMatrix class is used for dense matrices.  A LAPACK implementation
of the multivector is coming as soon as some conflicts between the NOX
and Teuchos BLAS/LAPACK wrappers are fixed.  Added a configure flag
(--enable-nox-multivectors) to enable/disable compiling in multivector
support as well as configure options for the location of the Teuchos include
and library directories.  Since LOCA needs multivectors directly, multivector
support is enabled if LOCA is enabled (the default), so to turn off
multivector support, you have to disable it and disable LOCA.  All of this
code compiles, but has not yet been tested for correctness, nor is there
any code that actually uses it yet.  The multivector interface is definitely
not set in stone and I encourage an open dialog on the developers list for
discussing it.

2)  Pointers in parameter lists:  Added an implementation of the
NOX::Parameter::Arbitrary class for passing pointers through parameter lists
using Teuchos::any (NOX::Parameter::AnyPtr).  Also added templated
set/get/is methods to the parameter list specific to pointers.  As for
multivector support, there is now a configure flag (--enable-nox-any) for
enabling (disabling) this feature, but is automatically enabled if LOCA
is enabled.  To see how pointers in parameter lists work, look in
 src-loca/example-lapack/Chan/ChanTPContinuation.C for an example of
adding pointers to parameters, and the top-most constructor in
src-loca/src/LOCA_Bifurcation_TPBord_ExtendedGroup.C for getting them
out.  I haven't commited the code yet that really makes use of this feature
however.

To add the Teuchos include path for compiling and library path for linking,
I had to add TRILINOS_CXXFLAGS and LIBS to most of the makefiles in the
repository.