Changes to make Mesquite compile on Dec and on HP (without the standard namespace and header files).  Changes are fairly pervasivie. 1)  Removing std:: prefixes from within the code.  Using a macro that adds 'using std::...' in the necessary files if needed.  The macro does nothing if the standard namespace isn't needed. 2)  Changing some #include statements so that the correct headers are included if USE_STD_INCLUDES is not defined for a given build. 3)  Adding #ifndef to make sure that PRINT_INFO and similar functions are only defined if they have not been previously definded. 4)  Changing some iterator comparisons to avoid using the '!=' comparison which is not defined in HP's non-ansi standard libraries.
(still needed:
In InstructionQueue.cpp two cout statement's have been commented
out to avoid a problem with string.  I'll try to fix this.
In MeshImpl.cpp, a file out behaves differently when USE_STD_INCLUDES is
not defined.  This problem should not affect the standard build.)

passes Regression and Unit on Sun.