----------------------------------------------------------------------
Committing in .

Added Files:
	Kokkos_Version.cpp
----------------------------------------------------------------------
Contributes to Bug 178

Kokkos_Version.cpp provides a function to return the version of Kokkos.

Other than a search and replace, the file is the same as Paul Sexton's
Tpetra_Version.cpp for Tpetra.

Having this file will allow me to autotool Kokkos (it is cleaner if there
is at least one .cpp file).