I found a way to add a arbitrary command after the installation target is finished.  What you do is to add a target called install-exec-hook and then you can put whatever commands you want in there (just make sure that they are portable).  I have used this feature to add the creation of a file called trilinos_make_options.mak in the install directory that contains make macros that are needed to define the compilers, linkers, libaries and other options that can be used by external makefiles to build and link against Trilinos.  I am using this in the MOOCHO makefiles.
This should not damage the portability of the generated makefile unless perl is not avaliable in the current path.  If perl is not available, then we have a problem and we will need to find a way to fix this.  However, since almost every system I have come across has perl and since the checkin testing script requires perl I do not see this to be a big problem.