Fixes for bug 811.
The Resize_ call has a check for whether it needs to run or not based
on the consistency of the existing size data.  This was breaking for
degenerate cases where the export count on some processors was zero.

Now, an explicit boolean is set to indicate that the resize operation
has already occurred.  Also, an all reduce operation is used to assure
all processors concur on their status.