To configure the Jpetra build tree and keep it compatible with the Trilinos specifications all source was moved into src/Jpetra.  When compiling java expects .java files to be in a directory that has the name of the package they belong to.  Hence the java source files need to reside in src/Jeptra.
All the files that support CCJ communication that don't fit the petra object module were moved to Jpetra/CcjSupport.
They reside in their own subpackage Jpetra.CcjSupport.

The Makefile in src/Jpetra is up-to-date and will compile Jpetra as long as the netlib blas and CCJ are
in the environmental class path.  A makejar command was added to the Makefile that generates a jar archive
in the jpetra package root named Jpetra.jar.  make all must be run first before make makejar can be run.

The makeJavaDocs.sh in doc has been updated to properly generated javadocs for Jpetra and its subpackages and place
that generated documentation into doc/html.

A new constructor was added to ElementSpace to support user defined linear contigous construction.

Export has been updated to determine how to distribute DistObjects which are linear.

GraphReader works correctly now.

CisMatrixWriter can send a CisMatrix to Matvis so that Matvis can use it/display it.

CisMatrixReader can read a CisMatrix from a url with a protocol that is supported by Java such as HTTP and FTP.